Uploading current Matlab App and ROS package
In this version it is possible to analyze graphs for the current fitness, average fitness, and for each of the 5 genes. On the App, it is still possible to select what type of environment pressure you want, set fixed genes, change mutation and neutralization, and save image and PDFs.
On the Gazebo simulation only six turtlebot3 robots are being used. (using turtlebot3 default map)